Casting away some integer conversion warnings that should not pose a problem.

+++ expat/examples/elements.c
@@ -52,7 +52,7 @@ main(int argc, char *argv[])
   XML_SetUserData(parser, &depth);
   XML_SetElementHandler(parser, startElement, endElement);
   do {
-    size_t len = fread(buf, 1, sizeof(buf), stdin);
+    int len = (int)fread(buf, 1, sizeof(buf), stdin);
     done = len < sizeof(buf);
     if (XML_Parse(parser, buf, len, done) == XML_STATUS_ERROR) {
       fprintf(stderr,
